Solve for "i" not defined AND adding in a "is_current" property for each page in Pages
So I'm finding this class useful. However when I implemented it, I found that I received an error stating that "i" was not defined. This simply defines the variable in the for loop.

Also, I thought it was a nice shortcut to add an "is_current" property to each page that references the "currentPage" property of the mixin.